Rename file to path
